v0.2.7 small fixes and optimizations, temp fix edit_order, clear button for unfocusing search field
