+++ lib/controllers/list/list_funcs.dart
