--- lib/theme/themes.dart
@@ -1,17 +1,0 @@
-import 'package:flutter/material.dart';
-
-ThemeData customDark = ThemeData(
-  colorScheme: ColorScheme.fromSeed(
-    seedColor: Colors.blue,
-    brightness: Brightness.dark,
-  ),
-  useMaterial3: true,
-);
-
-ThemeData customLight = ThemeData(
-  colorScheme: ColorScheme.fromSeed(
-    seedColor: Colors.blue,
-    brightness: Brightness.light,
-  ),
-  useMaterial3: true,
-);
