+++ lib/views/home/home_page_widgets.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 
-// Class for the product card in small form
 class ProductCard extends StatelessWidget {
   final String id;
   final String title;
@@ -25,6 +24,8 @@ class ProductCard extends StatelessWidget {
         borderRadius: BorderRadius.circular(20.0),
         child: InkWell(
           onTap: () {
+            // Unfocus before navigating
+            FocusScope.of(context).unfocus();
             Navigator.push(
               context,
               MaterialPageRoute(
@@ -54,15 +55,31 @@ class ProductCard extends StatelessWidget {
                 child: Column(
                   crossAxisAlignment: CrossAxisAlignment.start,
                   children: [
-                    Text(title,
-                        style: TextStyle(
-                            fontSize: 16.0, fontWeight: FontWeight.bold)),
+                    Row(
+                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                      children: [
+                        Expanded(
+                          child: Text(
+                            title,
+                            style: TextStyle(
+                                fontSize: 16.0, fontWeight: FontWeight.bold),
+                            overflow: TextOverflow.ellipsis,
+                          ),
+                        ),
+                        Text(
+                          price,
+                          style: TextStyle(
+                              fontSize: 16.0, fontWeight: FontWeight.bold),
+                        ),
+                      ],
+                    ),
                     SizedBox(height: 4.0),
-                    Text(description, style: TextStyle(fontSize: 12.0)),
-                    SizedBox(height: 8.0),
-                    Text(price,
-                        style: TextStyle(
-                            fontSize: 16.0, fontWeight: FontWeight.bold)),
+                    Text(
+                      description,
+                      style: TextStyle(fontSize: 12.0),
+                      maxLines: 2,
+                      overflow: TextOverflow.ellipsis,
+                    ),
                   ],
                 ),
               ),
@@ -74,7 +91,6 @@ class ProductCard extends StatelessWidget {
   }
 }
 
-// Class for the product detail page
 class ProductDetailPage extends StatelessWidget {
   final String id;
   final String title;
@@ -96,22 +112,21 @@ class ProductDetailPage extends StatelessWidget {
       appBar: AppBar(
         title: Text(title),
       ),
-      body: Padding(
-        padding: const EdgeInsets.all(16.0),
-        child: Center(
-          child: InkWell(
-            onTap: () => Navigator.pop(context),
-            child: Hero(
-              tag: 'product-$id',
-              child: Material(
-                elevation: 4.0,
-                borderRadius: BorderRadius.circular(10.0),
-                child: Column(
-                  mainAxisSize: MainAxisSize.min,
-                  children: [
-                    ClipRRect(
-                      borderRadius:
-                          BorderRadius.vertical(top: Radius.circular(20.0)),
+      body: GestureDetector(
+        onTap: () => Navigator.pop(context),
+        child: SingleChildScrollView(
+          child: Padding(
+            padding: const EdgeInsets.all(16.0),
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                Hero(
+                  tag: 'product-$id',
+                  child: Material(
+                    elevation: 4.0,
+                    borderRadius: BorderRadius.circular(20.0),
+                    child: ClipRRect(
+                      borderRadius: BorderRadius.circular(20.0),
                       child: Image.asset(
                         imagePath,
                         height: 300.0,
@@ -119,17 +134,99 @@ class ProductDetailPage extends StatelessWidget {
                         fit: BoxFit.cover,
                       ),
                     ),
-                    ListTile(
-                      title: Text(title),
-                      subtitle: Text(description),
-                      trailing: Text(price, style: TextStyle(fontSize: 20.0)),
+                  ),
+                ),
+                SizedBox(height: 16.0),
+                Row(
+                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                  children: [
+                    Expanded(
+                      child: Text(
+                        title,
+                        style: TextStyle(
+                            fontSize: 24.0, fontWeight: FontWeight.bold),
+                      ),
+                    ),
+                    Text(
+                      price,
+                      style: TextStyle(
+                          fontSize: 20.0, fontWeight: FontWeight.bold),
                     ),
                   ],
                 ),
-              ),
+                SizedBox(height: 8.0),
+                Text(
+                  description,
+                  style: TextStyle(fontSize: 16.0),
+                ),
+              ],
             ),
           ),
         ),
+      ),
+    );
+  }
+}
+
+class SearchField extends StatefulWidget {
+  final Function(String) onSearchChanged;
+  final String hintText;
+
+  SearchField({required this.onSearchChanged, required this.hintText});
+
+  @override
+  SearchFieldState createState() => SearchFieldState();
+}
+
+class SearchFieldState extends State<SearchField> {
+  final FocusNode _focusNode = FocusNode();
+  final TextEditingController _controller = TextEditingController();
+
+  @override
+  void initState() {
+    super.initState();
+    _focusNode.addListener(_onFocusChange);
+  }
+
+  @override
+  void dispose() {
+    _focusNode.removeListener(_onFocusChange);
+    _focusNode.dispose();
+    _controller.dispose();
+    super.dispose();
+  }
+
+  void _onFocusChange() {
+    // Only handle focus loss logic if needed
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.all(8.0),
+      child: TextField(
+        controller: _controller,
+        focusNode: _focusNode,
+        decoration: InputDecoration(
+          hintText: widget.hintText,
+          prefixIcon: Icon(Icons.search),
+          border: OutlineInputBorder(
+            borderRadius: BorderRadius.circular(20.0),
+          ),
+          suffixIcon: IconButton(
+            icon: Icon(Icons.clear),
+            onPressed: () {
+              _controller.clear();
+              widget.onSearchChanged('');
+              _focusNode.unfocus(); // Manually unfocus the search field
+            },
+          ),
+        ),
+        onChanged: (value) {
+          setState(
+              () {}); 
+          widget.onSearchChanged(value);
+        },
       ),
     );
   }
